+++ command/command.go
@@ -11,10 +11,12 @@ import (
 	"time"
 
 	"github.com/mackerelio/mackerel-agent/agent"
+	"github.com/mackerelio/mackerel-agent/checks"
 	"github.com/mackerelio/mackerel-agent/config"
 	"github.com/mackerelio/mackerel-agent/logging"
 	"github.com/mackerelio/mackerel-agent/mackerel"
 	"github.com/mackerelio/mackerel-agent/spec"
+	"github.com/mackerelio/mackerel-agent/util"
 )
 
 var logger = logging.GetLogger("command")
@@ -54,19 +56,19 @@ func saveHostID(root string, id string) error {
 }
 
 // buildHostSpec build data structure for Host specs
-func buildHostSpec(name string, meta map[string]interface{}, interfaces []map[string]interface{}, roleFullnames []string) map[string]interface{} {
-
+func buildHostSpec(name string, meta map[string]interface{}, interfaces []map[string]interface{}, roleFullnames []string, checks []string) map[string]interface{} {
 	return map[string]interface{}{
 		"name":          name,
 		"meta":          meta,
 		"interfaces":    interfaces,
 		"roleFullnames": roleFullnames,
+		"checks":        checks,
 	}
 }
 
 // prepareHost collects specs of the host and sends them to Mackerel server.
 // A unique host-id is returned by the server if one is not specified.
-func prepareHost(root string, api *mackerel.API, roleFullnames []string) (*mackerel.Host, error) {
+func prepareHost(root string, api *mackerel.API, roleFullnames []string, checks []string) (*mackerel.Host, error) {
 	// XXX this configuration should be moved to under spec/linux
 	os.Setenv("PATH", "/sbin:/usr/sbin:/bin:/usr/bin:"+os.Getenv("PATH"))
 	os.Setenv("LANG", "C") // prevent changing outputs of some command, e.g. ifconfig.
@@ -93,7 +95,7 @@ func prepareHost(root string, api *mackerel.API, roleFullnames []string) (*macke
 		if err != nil {
 			return nil, fmt.Errorf("Failed to find this host on mackerel (You may want to delete file \"%s\" to register this host to an another organization): %s", idFilePath(root), err.Error())
 		}
-		err := api.UpdateHost(hostID, buildHostSpec(hostname, meta, interfaces, roleFullnames))
+		err := api.UpdateHost(hostID, buildHostSpec(hostname, meta, interfaces, roleFullnames, checks))
 		if err != nil {
 			return nil, fmt.Errorf("Failed to update this host: %s", err.Error())
 		}
@@ -161,10 +163,23 @@ func loop(c *context, termCh chan struct{}) int {
 		c.ag.InitPluginGenerators(c.api)
 	}
 
+	termCheckerCh := make(chan struct{})
+	termMetricsCh := make(chan struct{})
+
+	// fan-out termCh
+	go func() {
+		for range termCh {
+			termCheckerCh <- struct{}{}
+			termMetricsCh <- struct{}{}
+		}
+	}()
+
+	runCheckersLoop(c, termCheckerCh, quit)
+
 	lState := loopStateFirst
 	for {
 		select {
-		case <-termCh:
+		case <-termMetricsCh:
 			if lState == loopStateTerminating {
 				return 1
 			}
@@ -217,7 +232,7 @@ func loop(c *context, termCh chan struct{}) int {
 			select {
 			case <-time.After(time.Duration(delaySeconds) * time.Second):
 				// nop
-			case <-termCh:
+			case <-termMetricsCh:
 				if lState == loopStateTerminating {
 					return 1
 				}
@@ -304,6 +319,113 @@ func enqueueLoop(c *context, postQueue chan *postValue, quit chan struct{}) {
 	}
 }
 
+// runCheckersLoop generates "checker" goroutines
+// which run for each checker commands and one for HTTP POSTing
+// the reports to Mackerel API.
+func runCheckersLoop(c *context, termCheckerCh <-chan struct{}, quit <-chan struct{}) {
+	var (
+		checkReportCh          chan *checks.Report
+		reportCheckImmediateCh chan struct{}
+	)
+	for _, checker := range c.ag.Checkers {
+		if checkReportCh == nil {
+			checkReportCh = make(chan *checks.Report)
+			reportCheckImmediateCh = make(chan struct{})
+		}
+
+		go func(checker checks.Checker) {
+			var (
+				lastStatus  = checks.StatusUndefined
+				lastMessage = ""
+			)
+
+			util.Periodically(
+				func() {
+					report, err := checker.Check()
+					if err != nil {
+						logger.Errorf("checker %v: %s", checker, err)
+						return
+					}
+
+					logger.Debugf("checker %q: report=%v", checker.Name, report)
+
+					if report.Status == lastStatus && report.Message == lastMessage {
+						// Do not report if nothing has changed
+						return
+					}
+
+					checkReportCh <- report
+
+					// If status has changed, send it immediately
+					// but if the status was OK and it's first invocation of a check, do not
+					if report.Status != lastStatus && !(report.Status == checks.StatusOK && lastStatus == checks.StatusUndefined) {
+						logger.Debugf("checker %q: status has changed %v -> %v: send it immediately", checker.Name, lastStatus, report.Status)
+						reportCheckImmediateCh <- struct{}{}
+					}
+
+					lastStatus = report.Status
+					lastMessage = report.Message
+				},
+				checker.Interval(),
+				quit,
+			)
+		}(checker)
+	}
+	if checkReportCh != nil {
+		go func() {
+			exit := false
+			for !exit {
+				select {
+				case <-time.After(1 * time.Minute):
+				case <-termCheckerCh:
+					logger.Debugf("received 'term' chan")
+					exit = true
+				case <-reportCheckImmediateCh:
+					logger.Debugf("received 'immediate' chan")
+				}
+
+				reports := []*checks.Report{}
+			DrainCheckReport:
+				for {
+					select {
+					case report := <-checkReportCh:
+						reports = append(reports, report)
+					default:
+						break DrainCheckReport
+					}
+				}
+
+				for i, report := range reports {
+					logger.Debugf("reports[%d]: %#v", i, report)
+				}
+
+				if len(reports) == 0 {
+					continue
+				}
+
+				err := c.api.ReportCheckMonitors(c.host.ID, reports)
+				if err != nil {
+					logger.Errorf("ReportCheckMonitors: %s", err)
+
+					// queue back the reports
+					go func() {
+						for _, report := range reports {
+							logger.Debugf("queue back report: %#v", report)
+							checkReportCh <- report
+						}
+					}()
+				}
+			}
+		}()
+	} else {
+		// consume termCheckerCh
+		go func() {
+			for range termCheckerCh {
+			}
+		}()
+	}
+}
+
 // collectHostSpecs collects host specs (correspond to "name", "meta" and "interfaces" fields in API v0)
 func collectHostSpecs() (string, map[string]interface{}, []map[string]interface{}, error) {
 	hostname, err := os.Hostname()
@@ -333,7 +455,7 @@ func UpdateHostSpecs(conf *config.Config, api *mackerel.API, host *mackerel.Host
 		return
 	}
 
-	err = api.UpdateHost(host.ID, buildHostSpec(hostname, meta, interfaces, conf.Roles))
+	err = api.UpdateHost(host.ID, buildHostSpec(hostname, meta, interfaces, conf.Roles, conf.CheckNames()))
 	if err != nil {
 		logger.Errorf("Error while updating host specs: %s", err)
 	} else {
@@ -349,7 +471,7 @@ func Prepare(conf *config.Config) (*mackerel.API, *mackerel.Host, error) {
 		return nil, nil, fmt.Errorf("Failed to prepare an api: %s", err.Error())
 	}
 
-	host, err := prepareHost(conf.Root, api, conf.Roles)
+	host, err := prepareHost(conf.Root, api, conf.Roles, conf.CheckNames())
 	if err != nil {
 		return nil, nil, fmt.Errorf("Failed to prepare host: %s", err.Error())
 	}
@@ -364,15 +486,12 @@ func RunOnce(conf *config.Config) {
 		logger.Errorf("While collecting host specs: %s", err)
 		return
 	}
-	ag := &agent.Agent{
-		MetricsGenerators: metricsGenerators(conf),
-		PluginGenerators:  pluginGenerators(conf),
-	}
+	ag := NewAgent(conf)
 	graphdefs := ag.CollectGraphDefsOfPlugins()
 	logger.Infof("Collecting metrics may take one minutes.")
 	metrics := ag.CollectMetrics(time.Now())
 	payload := map[string]interface{}{
-		"host":    buildHostSpec(hostname, meta, interfaces, conf.Roles),
+		"host":    buildHostSpec(hostname, meta, interfaces, conf.Roles, conf.CheckNames()),
 		"metrics": metrics,
 	}
 	json, err := json.Marshal(payload)
@@ -383,14 +502,20 @@ func RunOnce(conf *config.Config) {
 	}
 }
 
+// NewAgent creates a new instance of agent.Agent from its configuration conf.
+func NewAgent(conf *config.Config) *agent.Agent {
+	return &agent.Agent{
+		MetricsGenerators: metricsGenerators(conf),
+		PluginGenerators:  pluginGenerators(conf),
+		Checkers:          createCheckers(conf),
+	}
+}
+
 // Run starts the main metric collecting logic and this function will never return.
 func Run(conf *config.Config, api *mackerel.API, host *mackerel.Host, termCh chan struct{}) int {
 	logger.Infof("Start: apibase = %s, hostName = %s, hostID = %s", conf.Apibase, host.Name, host.ID)
 
-	ag := &agent.Agent{
-		MetricsGenerators: metricsGenerators(conf),
-		PluginGenerators:  pluginGenerators(conf),
-	}
+	ag := NewAgent(conf)
 
 	c := &context{
 		ag:   ag,
@@ -400,4 +525,19 @@ func Run(conf *config.Config, api *mackerel.API, host *mackerel.Host, termCh cha
 	}
 
 	return loop(c, termCh)
+}
+
+func createCheckers(conf *config.Config) []checks.Checker {
+	checkers := []checks.Checker{}
+
+	for name, pluginConfig := range conf.Plugin["checks"] {
+		checker := checks.Checker{
+			Name:   name,
+			Config: pluginConfig,
+		}
+		logger.Debugf("Checker created: %v", checker)
+		checkers = append(checkers, checker)
+	}
+
+	return checkers
 }