+++ metrics/windows/memory.go
@@ -36,8 +36,8 @@ func (g *MemoryGenerator) Generate() (metrics.Values, error) {
 	ret["memory.free"] = free
 	ret["memory.total"] = total
 	ret["memory.used"] = total - free
-	ret["memory.swap_total"] = float64(memoryStatusEx.TotalVirtual) / 1024
-	ret["memory.swap_free"] = float64(memoryStatusEx.AvailVirtual) / 1024
+	ret["memory.pagefile_total"] = float64(memoryStatusEx.TotalPageFile) / 1024
+	ret["memory.pagefile_free"] = float64(memoryStatusEx.AvailPageFile) / 1024
 
 	memoryLogger.Debugf("memory : %s", ret)
 	return metrics.Values(ret), nil