+++ util/util.go
@@ -16,8 +16,8 @@ var timeoutDuration = 30 * time.Second
 // timeoutKillAfter is option of `RunCommand()` set waiting limit to `kill -kill` after terminating the command.
 var timeoutKillAfter = 10 * time.Second
 
-// RunCommand runs command (in one string) and returns stdout, stderr strings.
-func RunCommand(command string) (string, string, error) {
+// RunCommand runs command (in one string) and returns stdout, stderr strings and its exit code.
+func RunCommand(command string) (string, string, int, error) {
 	tio := &timeout.Timeout{
 		Cmd:       exec.Command("/bin/sh", "-c", command),
 		Duration:  timeoutDuration,
@@ -28,5 +28,6 @@ func RunCommand(command string) (string, string, error) {
 	if err == nil && exitStatus.IsTimedOut() {
 		err = fmt.Errorf("command timed out")
 	}
-	return stdout, stderr, err
+
+	return stdout, stderr, exitStatus.GetChildExitCode(), err
 }